Изменено поле user_id сущностей для оценки с id на email

+++ app/src/main/java/com/company/metrix/auth/estimates/FragmentEstimate.kt
@@ -120,7 +120,7 @@ class FragmentEstimate() : Fragment() {
     }
 
     private fun onSubmit() {
-        val userId = binding.employeeIdField.text.toString().trim().toLong()
+        val userId = binding.employeeIdField.text.toString().trim()
         userRating = binding.ratingBar.rating.toDouble()
         userComment = binding.employeeCommentField.text.toString().trim()
         viewModel.viewModelScope.launch {
